Rename vocalpy.vendor -> vocalpy._vendor
