+++ src/vocalpy/_vendor/__init__.py
@@ -1,0 +1,1 @@
+from . import evfuncs
